pyproject.toml
[project]
name = "domainraptor"
version = "0.2.0"
description = "Cyber Intelligence tool for domain reconnaissance and vulnerability assessment"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.12"
keywords = ["security", "reconnaissance", "osint", "domain", "vulnerability"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Information Technology",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Security",
    "Topic :: Internet :: Name Service (DNS)",
    "Typing :: Typed",
]

dependencies = [
    "shodan>=1.31.0",
    "colorama>=0.4.6",
    "pygments>=2.18.0",
    "python-whois>=0.9.4",
    "requests>=2.32.0",
    "pyyaml>=6.0.1",
    "dnspython>=2.6.1",
    "sslyze>=6.0.0",
    "httpx>=0.27.0",
    "rich>=13.7.0",
    "typer>=0.12.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.5.0",
    "pre-commit>=3.7.0",
    "pytest>=8.2.0",
    "pytest-cov>=5.0.0",
    "pytest-asyncio>=0.23.0",
    "mypy>=1.10.0",
    "types-requests>=2.32.0",
    "types-PyYAML>=6.0.0",
]
docs = ["mkdocs>=1.6.0", "mkdocs-material>=9.5.0"]

[project.scripts]
domainraptor = "domainraptor.cli:main"
dr = "domainraptor.cli:main"

[project.urls]
Homepage = "https://github.com/yourusername/DomainRaptor"
Documentation = "https://github.com/yourusername/DomainRaptor#readme"
Repository = "https://github.com/yourusername/DomainRaptor.git"
Issues = "https://github.com/yourusername/DomainRaptor/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.sdist]
exclude = ["/.github", "/docs", "/tests"]

[tool.hatch.build.targets.wheel]
packages = ["src/domainraptor"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = ["-v", "--tb=short", "--strict-markers"]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests requiring external APIs",
]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
strict_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
show_error_codes = true

[[tool.mypy.overrides]]
module = ["shodan.*", "whois.*", "sslyze.*"]
ignore_missing_imports = true

[tool.coverage.run]
source = ["src/domainraptor"]
branch = true
omit = ["tests/*", "*/__pycache__/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
    "if __name__ == .__main__.:",
]
fail_under = 70

[tool.bandit]
exclude_dirs = ["tests", "venv", ".venv"]
skips = ["B101", "B311"]                  # B101: assert, B311: pseudo-random (ok for non-crypto)
targets = ["."]

[tool.bandit.assert_used]
skips = ["*_test.py", "*test_*.py"]
